requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.9
# by the following command:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
--index-url https://pypi.tuna.tsinghua.edu.cn/simple
--extra-index-url https://mirrors.aliyun.com/pypi/simple
--extra-index-url https://pypi.doubanio.com/simple
--extra-index-url https://pypi.org/simple

aiohttp==3.8.4
    # via fsspec
aiosignal==1.3.1
    # via aiohttp
async-timeout==4.0.2
    # via aiohttp
attrs==23.1.0
    # via aiohttp
certifi==2023.5.7
    # via requests
charset-normalizer==3.1.0
    # via
    #   aiohttp
    #   requests
click==8.1.3
    # via
    #   -r requirements.in
    #   cnstd
cnstd==1.2.3
    # via -r requirements.in
coloredlogs==15.0.1
    # via onnxruntime
contourpy==1.1.0
    # via matplotlib
cycler==0.11.0
    # via matplotlib
filelock==3.12.2
    # via
    #   huggingface-hub
    #   torch
flatbuffers==23.5.26
    # via onnxruntime
fonttools==4.40.0
    # via matplotlib
frozenlist==1.3.3
    # via
    #   aiohttp
    #   aiosignal
fsspec[http]==2023.6.0
    # via
    #   huggingface-hub
    #   pytorch-lightning
huggingface-hub==0.15.1
    # via cnstd
humanfriendly==10.0
    # via coloredlogs
idna==3.4
    # via
    #   requests
    #   yarl
importlib-resources==5.12.0
    # via matplotlib
jinja2==3.1.2
    # via torch
kiwisolver==1.4.4
    # via matplotlib
lightning-utilities==0.9.0
    # via pytorch-lightning
markupsafe==2.1.3
    # via jinja2
matplotlib==3.7.1
    # via
    #   cnstd
    #   seaborn
mpmath==1.3.0
    # via sympy
multidict==6.0.4
    # via
    #   aiohttp
    #   yarl
networkx==3.1
    # via torch
numpy==1.25.0
    # via
    #   -r requirements.in
    #   cnstd
    #   contourpy
    #   matplotlib
    #   onnx
    #   onnxruntime
    #   opencv-python
    #   pandas
    #   pytorch-lightning
    #   scipy
    #   seaborn
    #   shapely
    #   torchmetrics
    #   torchvision
onnx==1.14.0
    # via
    #   -r requirements.in
    #   cnstd
onnxruntime==1.15.1
    # via
    #   -r requirements.in
    #   cnstd
opencv-python==4.7.0.72
    # via cnstd
packaging==23.1
    # via
    #   huggingface-hub
    #   lightning-utilities
    #   matplotlib
    #   onnxruntime
    #   pytorch-lightning
    #   torchmetrics
pandas==2.0.3
    # via
    #   cnstd
    #   seaborn
pillow==9.5.0
    # via
    #   -r requirements.in
    #   cnstd
    #   matplotlib
    #   torchvision
qrcode==7.4.2
polygon3==3.0.9.1
    # via cnstd
protobuf==4.23.3
    # via
    #   onnx
    #   onnxruntime
pyclipper==1.3.0.post4
    # via cnstd
pyparsing==3.1.0
    # via matplotlib
python-dateutil==2.8.2
    # via
    #   matplotlib
    #   pandas
pytorch-lightning==2.0.4
    # via
    #   -r requirements.in
    #   cnstd
pytz==2023.3
    # via pandas
pyyaml==6.0
    # via
    #   cnstd
    #   huggingface-hub
    #   pytorch-lightning
requests==2.31.0
    # via
    #   fsspec
    #   huggingface-hub
    #   torchvision
scipy==1.11.1
    # via cnstd
seaborn==0.12.2
    # via cnstd
shapely==2.0.1
    # via cnstd
six==1.16.0
    # via python-dateutil
sympy==1.12
    # via
    #   onnxruntime
    #   torch
torch==2.0.1
    # via
    #   -r requirements.in
    #   cnstd
    #   pytorch-lightning
    #   torchmetrics
    #   torchvision
torchmetrics==0.11.4
    # via
    #   -r requirements.in
    #   pytorch-lightning
torchvision==0.15.2
    # via
    #   -r requirements.in
    #   cnstd
tqdm==4.65.0
    # via
    #   -r requirements.in
    #   cnstd
    #   huggingface-hub
    #   pytorch-lightning
typing-extensions==4.7.0
    # via
    #   huggingface-hub
    #   lightning-utilities
    #   onnx
    #   pytorch-lightning
    #   torch
tzdata==2023.3
    # via pandas
unidecode==1.3.6
    # via cnstd
urllib3==2.0.3
    # via requests
yarl==1.9.2
    # via aiohttp
zipp==3.15.0
    # via importlib-resources
mysql-connector-python~=8.1.0
